refactor: consolidate backup package into single module
Combine service, metadata, helpers and restore into a single
src/my_unicorn/core/backup.py. Remove former files and update tests
to import from my_unicorn.core.backup.

Closes #278
